+++ docs/CHANGE_LOG.md
@@ -1,7 +1,11 @@
 # Heat.js - Change Log:
 
 ## Version 1.5.0:
-- 
+
+#### **Binding Options - Custom Triggers:**
+- Added a new binding option called "onAdd", which states an event that should be called when new dates are added.
+- Added a new binding option called "onRemove", which states an event that should be called when dates are removed.
+- Added a new binding option called "onReset", which states an event that should be called when all data for an element is reset.
 
 <br>